Extract `internal/fileurl` package
